requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0
celery==5.3.6
requests==2.31.0
reportlab==4.1.0
pillow==10.3.0
python-decouple==3.8
django-cors-headers==4.3.1
openpyxl==3.1.2
phonenumbers==8.13.37
django-filter==24.2
drf-spectacular==0.27.2

gunicorn==22.0.0
python-dateutil==2.9.0.post0
django-celery-beat==2.6.0
